Renamed all references to old name to new
