+++ eICU_Example/pre_MEDS.py
@@ -18,11 +18,7 @@ import polars as pl
 from loguru import logger
 from omegaconf import DictConfig, OmegaConf
 
-from MEDS_polars_functions.utils import (
-    get_shard_prefix,
-    hydra_loguru_init,
-    write_lazyframe,
-)
+from MEDS_transforms.utils import get_shard_prefix, hydra_loguru_init, write_lazyframe
 
 HEALTH_SYSTEM_STAY_ID = "patienthealthsystemstayid"
 UNIT_STAY_ID = "patientunitstayid"
