+++ src/MEDS_transforms/configs/stage_configs/fit_vocabulary_indices.yaml
