+++ src/MEDS_transforms/extract/__init__.py
@@ -1,4 +1,4 @@
-from MEDS_polars_functions import EXTRACT_CONFIG_YAML
+from MEDS_transforms import EXTRACT_CONFIG_YAML
 
 # We set this equality explicitly here so linting does not remove an apparently "unused" import if we just
 # rename with "as" during the import.
