+++ src/MEDS_transforms/extract/shard_events.py
@@ -13,9 +13,9 @@ import polars as pl
 from loguru import logger
 from omegaconf import DictConfig, OmegaConf
 
-from MEDS_polars_functions.extract import CONFIG_YAML
-from MEDS_polars_functions.mapreduce.mapper import rwlock_wrap
-from MEDS_polars_functions.utils import (
+from MEDS_transforms.extract import CONFIG_YAML
+from MEDS_transforms.mapreduce.mapper import rwlock_wrap
+from MEDS_transforms.utils import (
     get_shard_prefix,
     hydra_loguru_init,
     is_col_field,
