+++ src/MEDS_transforms/transforms/add_time_derived_measurements.py
@@ -7,9 +7,9 @@ import polars as pl
 from loguru import logger
 from omegaconf import DictConfig, OmegaConf
 
-from MEDS_polars_functions import PREPROCESS_CONFIG_YAML
-from MEDS_polars_functions.mapreduce.mapper import map_over
-from MEDS_polars_functions.utils import hydra_loguru_init
+from MEDS_transforms import PREPROCESS_CONFIG_YAML
+from MEDS_transforms.mapreduce.mapper import map_over
+from MEDS_transforms.utils import hydra_loguru_init
 
 pl.enable_string_cache()
