+++ src/MEDS_transforms/transforms/normalization.py
@@ -8,8 +8,8 @@ import hydra
 import polars as pl
 from omegaconf import DictConfig
 
-from MEDS_polars_functions import PREPROCESS_CONFIG_YAML
-from MEDS_polars_functions.mapreduce.mapper import map_over
+from MEDS_transforms import PREPROCESS_CONFIG_YAML
+from MEDS_transforms.mapreduce.mapper import map_over
 
 
 def normalize(
