+++ src/MEDS_transforms/transforms/occlude_outliers.py
@@ -7,8 +7,8 @@ import hydra
 import polars as pl
 from omegaconf import DictConfig
 
-from MEDS_polars_functions import PREPROCESS_CONFIG_YAML
-from MEDS_polars_functions.mapreduce.mapper import map_over
+from MEDS_transforms import PREPROCESS_CONFIG_YAML
+from MEDS_transforms.mapreduce.mapper import map_over
 
 pl.enable_string_cache()
